requirements.txt
gensim==4.3.2
matplotlib==3.5.2
nltk==3.8.1
numpy==1.22.4
pandas==2.0.3
pyenchant==3.2.2
scikit_learn==1.3.2
torch==1.11.0+cu113
transformers==4.36.0
wget==3.2
